a6.5.1

+++ alpha/MyBox/src/main/java/mara/mybox/tools/DoubleMatrixTools.java
@@ -44,20 +44,20 @@ public class DoubleMatrixTools {
         return rv;
     }
 
-    public static float[] matrix2Array(float[][] m) {
+    public static double[] matrix2Array(double[][] m) {
         if (m == null || m.length == 0 || m[0].length == 0) {
             return null;
         }
         int h = m.length;
         int w = m[0].length;
-        float[] a = new float[w * h];
+        double[] a = new double[w * h];
         for (int j = 0; j < h; ++j) {
             System.arraycopy(m[j], 0, a, j * w, w);
         }
         return a;
     }
 
-    public static float[][] array2Matrix(float[] a, int w) {
+    public static double[][] array2Matrix(double[] a, int w) {
         if (a == null || a.length == 0 || w < 1) {
             return null;
         }
@@ -65,7 +65,7 @@ public class DoubleMatrixTools {
         if (h < 1) {
             return null;
         }
-        float[][] m = new float[h][w];
+        double[][] m = new double[h][w];
         for (int j = 0; j < h; ++j) {
             for (int i = 0; i < w; ++i) {
                 m[j][i] = a[j * w + i];
